requirements.txt
Cython==0.23.4
ipdb==0.8.1
ipython==4.0.0
mysqlclient==1.3.6
nose==1.3.7
coverage==4.0.2
numpy==1.11.0
pandas==0.18.0
PuLP==1.6.0
pyyaml==3.11.0
unidecode==0.4.19
scrapy==1.1.0rc3
